pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spam-eggs"
version = "2020.0.0"
requires-python = ">=3.8"
dependencies = [
	"colour>=0.1.5",
    "matplotlib>=3.7.1",
    "numpy>=1.26.4",
    "Pillow>=9.4.0",
    "Pillow>=10.2.0",
    "scikit_learn>=1.3.0",
    "setuptools>=68.0.0",
    "tqdm>=4.64.0",
]
authors = [
  {name = "Andrea Combette", email = "pradyun@example.com"},
]
maintainers = [
  {name = "Andrea Combette", email = "brett@example.com"}
]
description = "Lovely Spam! Wonderful Spam!"
readme = "README.rst"
license = {file = "LICENSE.txt"}
classifiers = [
  "Programming Language :: Python",
]
[project.urls]
Homepage = "https://example.com"
Documentation = "https://readthedocs.org"
Repository = "https://github.com/me/spam.git"
"Bug Tracker" = "https://github.com/me/spam/issues"
Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[project.scripts]
theme-changer = "main:main"
